Offload siwe to its own composable
